add --rename option to rename declarations and references
The `--rename` option takes a string as input, treating every couple, separated by a space,
as a find-replace option.
For example,  `--rename "$theService $myService "theFactory myFactory"`will replace
`$theService` with `$myService` and `theFactory` with `myFactory` in both the declaration
and in any annotation that will be added by ng-annotate. The actual use of the provider
won't be changed.

Also, a couple of JS grammar fixes here and there.

Closes #44